Refactors the next style method into speech generator.
