+++ ts/speech_generator/abstract_speech_generator.ts
@@ -26,6 +26,7 @@ import { RebuildStree } from '../walker/rebuild_stree.js';
 import { SpeechGenerator } from './speech_generator.js';
 import * as SpeechGeneratorUtil from './speech_generator_util.js';
 import * as EngineConst from '../common/engine_const.js';
+import { SemanticNode } from '../semantic_tree/semantic_node.js';
 
 import { ClearspeakPreferences } from '../speech_rules/clearspeak_preferences.js';
 
@@ -120,6 +121,54 @@ export abstract class AbstractSpeechGenerator implements SpeechGenerator {
   /**
    * @override
    */
-  public nextStyle() { }
+  public nextStyle(id: string) {
+    this.setOption('style', this.nextStyle_(this.getRebuilt().nodeDict[id]));
+  }
+  
+  /**
+   * Cycles to next style or preference of the speech rule set if possible.
+   *
+   * @param node The semantic node currently in focus.
+   * @returns The new style name.
+   */
+  private nextStyle_(node: SemanticNode): string {
+    const {modality: modality, domain: domain, style: style} = this.getOptions();
+    // Rule cycling only makes sense for speech modality.
+    if (modality !== 'speech') {
+      return style;
+    }
+
+    if (domain === 'mathspeak') {
+      const styles = ['default', 'brief', 'sbrief'];
+      const index = styles.indexOf(style);
+      if (index === -1) {
+        return style;
+      }
+      return index >= styles.length - 1 ? styles[0] : styles[index + 1];
+    }
+    if (domain === 'clearspeak') {
+      const prefs = ClearspeakPreferences.getLocalePreferences();
+      const loc = prefs['en'];
+      // TODO: use correct locale.
+      if (!loc) {
+        return 'default';
+      }
+      // TODO: return the previous one?
+      const smart = ClearspeakPreferences.relevantPreferences(node);
+      const current = ClearspeakPreferences.findPreference(style, smart);
+      const options = loc[smart].map(function (x) {
+        return x.split('_')[1];
+      });
+      const index = options.indexOf(current);
+      if (index === -1) {
+        return style;
+      }
+      const next =
+        index >= options.length - 1 ? options[0] : options[index + 1];
+      const result = ClearspeakPreferences.addPreference(style, smart, next);
+      return result;
+    }
+    return style;
+  }
 
 }
